docker/civitas-dummy-generator/requirements.txt
# Core Scheduling & Framework
APScheduler==3.10.4
click==8.1.7
pydantic==2.5.2
pydantic-settings==2.1.0

# Data Generation & Processing
Faker==22.0.0
pandas==2.1.4
numpy==1.26.2

# Database & Storage Connectivity
psycopg2-binary==2.9.9
boto3==1.34.34
paho-mqtt==1.6.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
